feat #23 - QuestionAndAnswer 모듈화 개선선
