+++ src/components/hiStory/CompltedEvents/QuestionAndAnswer.tsx
@@ -3,20 +3,22 @@ import { useEffect, useState } from 'react';
 
 interface Props {
   question: QuestionContentType | CompleteEventType['representativeQuestion'];
+  title?: string;
   isRepresent?: boolean;
   isAvailBookmark?: boolean;
 }
 
 const QuestionAndAnswer = ({
   question,
+  title,
   isRepresent = false,
   isAvailBookmark = false,
 }: Props) => {
   const [showAnswer, setShowAnswer] = useState(false);
 
   useEffect(() => {
-    setShowAnswer(false);
-  }, [question]);
+    setShowAnswer(!!title);
+  }, [question, title]);
 
   const backgroundStyle = isRepresent
     ? {
@@ -42,7 +44,6 @@ const QuestionAndAnswer = ({
           />
         );
       case 'VOICE':
-        // TODO: Implement voice answer rendering
         return (
           <p className="leading-[24px]">Voice answer not implemented yet</p>
         );
@@ -52,39 +53,51 @@ const QuestionAndAnswer = ({
   };
 
   return (
-    <div
-      className={`w-full h-auto py-4 mb-2.5 self-center rounded-[10px]`}
-      style={backgroundStyle}
-      onClick={() => setShowAnswer(!showAnswer)}
-    >
-      <div className="w-full flex flex-col items-start justify-center px-[10px]">
-        <div className="w-full flex">
-          <div className="flex-shrink-0 w-[34px]">
-            <span className="font-pretendard-300 text-[#BAD7EC] text-[24px] leading-[24px]">
-              Q.
+    <div className="w-[calc(100%-40px)] mx-auto">
+      {title && (
+        <div className="w-full flex justify-center">
+          <div className="w-[calc(100%-24px)] h-[20px] mb-3">
+            <span className="font-pretendard-400 text-[#888] text-[14px]">
+              {title}
             </span>
           </div>
-          <span className="font-pretendard-200 text-[14px] flex-grow leading-[24px]">
-            {question.content}
-          </span>
         </div>
+      )}
 
-        {question.isAnswer && showAnswer && (
-          <div className="w-full flex mt-2">
+      <div
+        className={`w-full h-auto py-4 mb-2.5 self-center rounded-[10px]`}
+        style={backgroundStyle}
+        onClick={() => !title && setShowAnswer(!showAnswer)}
+      >
+        <div className="w-full flex flex-col items-start justify-center px-[10px]">
+          <div className="w-full flex">
             <div className="flex-shrink-0 w-[34px]">
               <span className="font-pretendard-300 text-[#BAD7EC] text-[24px] leading-[24px]">
-                A.
+                Q.
               </span>
             </div>
-            <div className="flex-grow font-pretendard-200 text-[14px]">
-              {renderAnswer(question.type!)}
-            </div>
+            <span className="font-pretendard-200 text-[14px] flex-grow leading-[24px]">
+              {question.content}
+            </span>
           </div>
-        )}
 
-        {/* {isBookmarking && (
+          {question.isAnswer && (showAnswer || !!title) && (
+            <div className="w-full flex mt-2">
+              <div className="flex-shrink-0 w-[34px]">
+                <span className="font-pretendard-300 text-[#BAD7EC] text-[24px] leading-[24px]">
+                  A.
+                </span>
+              </div>
+              <div className="flex-grow font-pretendard-200 text-[14px]">
+                {renderAnswer(question.type!)}
+              </div>
+            </div>
+          )}
+
+          {/* {isBookmarking && (
           // TODO: Implement bookmarking functionality
         )} */}
+        </div>
       </div>
     </div>
   );
