+++ src/types/emotion.ts
@@ -1,0 +1,7 @@
+export type EmotionType =
+  | 'JOY'
+  | 'HAPPINESS'
+  | 'LOVE'
+  | 'SATISFACTION'
+  | 'HOPE'
+  | 'EXPECTATION';
